Implementing Redux in BH
